+++ src/components/10-atoms/input-text/react/demo.jsx
@@ -4,6 +4,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import DemoInputText from './DemoInputText';
 import Changelog from '../CHANGELOG.md';
+import DemoInputTextOnInputfieldKeyUp
+  from "../../../10-atoms/input-text/react/DemoInputTextOnInputfieldKeyUp";
 
 storiesOf('Examples/Input Text/React', module)
   .addParameters({
@@ -13,5 +15,10 @@ storiesOf('Examples/Input Text/React', module)
   .add('Controlled/uncontrolled', () => {
     const div = document.createElement('div');
     ReactDOM.render(<DemoInputText />, div);
+    return div;
+  })
+  .add('Using onInputfieldKeyUp event', () => {
+    const div = document.createElement('div');
+    ReactDOM.render(<DemoInputTextOnInputfieldKeyUp />, div);
     return div;
   });